backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["mlbot321 <example@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
scikit-learn = "^1.6.0"
numpy = "^1.26.4"
sentence-transformers = "^3.3.1"
faiss-cpu = "^1.9.0.post1"
pandas = "^2.2.3"
nbstripout = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
